healthcare_agent/requirements.txt
# 核心框架
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LangChain + LlamaIndex
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
llama-index>=0.9.0
llama-index-core>=0.9.0
llama-index-embeddings-langchain>=0.1.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-readers-file>=0.1.0

# 向量数据库
chromadb>=0.4.0

# 模型推理
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# 阿里云API
httpx>=0.26.0
dashscope>=1.14.0

# 文档处理
pymupdf>=1.23.0
pdfplumber>=0.10.0
Pillow>=10.0.0
pytesseract>=0.3.10
markdown>=3.5.1

# 中文分词
jieba>=0.42.1
pyahocorasick>=1.5.3

# 定时任务
apscheduler>=3.10.0

# 数据库
sqlalchemy>=2.0.0

# 工具库
numpy>=1.24.0
pyyaml>=6.0
cryptography>=41.0.0

# 企微SDK
wechatpy>=1.8.0

# OCR
paddleocr>=2.7.0
paddlepaddle>=2.5.0
